refactor: split oauth.ts into logical modules
Split the 850-line oauth.ts into smaller, focused modules:
- types.ts: Interfaces (OAuthConfig, OAuthCallbackResult) and constants
- storage.ts: File storage utilities for tokens, clients, verifiers
- browser.ts: Cross-platform browser opening utility
- callback-server.ts: HTTP callback server with HTML templates
- provider.ts: Main McpCliOAuthProvider class
- index.ts: Re-exports for backwards compatibility
